backend/requirements.txt
Flask
flask-socketio
flask_cors
python-dotenv